Moved gallery business logic from gallery service to gallery view model
