+++ client/lib/features/gallery/services/gallery_service.dart
@@ -3,26 +3,9 @@ import 'dart:io';
 import 'package:georeal/constants/env_variables.dart';
 import 'package:http/http.dart' as http;
 
-import '../../../models/gallery_model.dart';
-
 /// Handles http requests for the Gallery
 
 class GalleryService {
-  final Map<String, Gallery> _galleries = {};
-
-  void createGalleryForGeoSphere(String geoSphereId) {
-    _galleries[geoSphereId] = Gallery(id: geoSphereId);
-  }
-
-  void addPhotoToGallery(String geoSphereId, String photoPath) {
-    final gallery = _galleries[geoSphereId];
-    if (gallery != null) {
-      gallery.photoPaths.add(photoPath);
-    } else {
-      print("Gallery with this id does not exist!");
-    }
-  }
-
   Future<void> uploadPhoto(String geoSphereId, File photo) async {
     var uri = Uri.parse('${EnvVariables.uri}/geofences');
 
@@ -41,9 +24,5 @@ class GalleryService {
     } catch (e) {
       throw Exception("Error occurred: $e");
     }
-  }
-
-  List<String> getPhotosFromGallery(String geoSphereId) {
-    return _galleries[geoSphereId]?.photoPaths ?? [];
   }
 }
