+++ client/lib/features/gallery/widgets/photo_prompt.dart
@@ -1,12 +1,12 @@
+import 'dart:developer';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:georeal/features/gallery/view_model/gallery_view_model.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:provider/provider.dart';
-
-import '../../gallery/services/gallery_service.dart';
 
 /// Prompt to add a photo to a GeoSphere
 
@@ -21,18 +21,18 @@ class PhotoPrompt extends StatefulWidget {
 class _PhotoPromptState extends State<PhotoPrompt> {
   File? image;
 
-  Future pickImage(ImageSource source) async {
+  Future pickImage(ImageSource source, BuildContext context) async {
     try {
       final image = await ImagePicker().pickImage(source: source);
       if (image == null) return;
 
       final imageTemporary = File(image.path);
       this.image = imageTemporary;
-
+      log("beofre");
       // Save the image path to the gallery
-      Provider.of<GalleryService>(context, listen: false)
+      Provider.of<GalleryViewModel>(context, listen: false)
           .addPhotoToGallery(widget.geoSphereId, image.path);
-
+      log("after");
       /*
       Provider.of<GalleryService>(context, listen: false)
           .uploadPhoto(widget.geoSphereId, imageTemporary);*/
@@ -50,7 +50,7 @@ class _PhotoPromptState extends State<PhotoPrompt> {
         child: Column(
           children: [
             ElevatedButton(
-              onPressed: () => pickImage(ImageSource.gallery),
+              onPressed: () => pickImage(ImageSource.gallery, context),
               child: const Text("Pick Gallery"),
             ),
             ElevatedButton(
@@ -58,13 +58,13 @@ class _PhotoPromptState extends State<PhotoPrompt> {
                 if (Platform.isAndroid) {
                   var status = await Permission.camera.request();
                   if (status.isGranted) {
-                    pickImage(ImageSource.camera);
+                    pickImage(ImageSource.camera, context);
                   } else {
                     print("bruh");
                   }
                 } else if (Platform.isIOS) {
                   // TODO: Implement iOS camera permission
-                  pickImage(ImageSource.camera);
+                  pickImage(ImageSource.camera, context);
                 }
               },
               child: const Text("Pick Camera"),
