+++ client/lib/home_router.dart
@@ -1,27 +1,7 @@
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
-import 'package:georeal/constants/env_variables.dart';
 import 'package:georeal/features/geo_sphere/views/geo_spheres_view.dart';
 import 'package:georeal/features/home/screens/home_screen.dart';
 import 'package:georeal/global_variables.dart';
-import 'package:http/http.dart' as http;
-
-Future<void> checkServerStatus() async {
-  try {
-    print("[georeal] Checking server");
-    final response = await http.get(Uri.parse(EnvVariables.uri));
-
-    if (response.statusCode == 200) {
-      var data = json.decode(response.body);
-      print('[georeal] Server status: ${data['status']}');
-    } else {
-      print('[georeal] Server error: ${response.statusCode}');
-    }
-  } catch (e) {
-    print('[georeal] Error checking server status: $e');
-  }
-}
 
 import 'features/geo_sphere/widgets/add_geo_sphere_modal.dart';
 
@@ -38,7 +18,6 @@ class _HomeRouterState extends State<HomeRouter> {
   @override
   void initState() {
     super.initState();
-    checkServerStatus();
   }
 
   int _page = 0;
